Crawler update - no args, 8 threads, working well

+++ Crawler/CrawlerCodes/Feeder.py
@@ -11,6 +11,10 @@ def feedCrawler():
 
     for row in results:
         alreadyStoredLinks.append(row[0])
+    
+    if len(alreadyStoredLinks) == 0:
+        print('No links to feed')
+        exit(-1)
 
     with open('Crawler\\CrawlData\\LinksToCrawl.json', 'w', encoding='utf-8') as outfile:
         json.dump(alreadyStoredLinks, outfile, indent=4, ensure_ascii=False)
